# swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "SWFIT CLOUD",
      "version": "1.0.0",
      "description": "Swift cloud api definations"
    },
    "servers": [
      {
        "url": "http://localhost:3000/"
      }
    ],
    "paths": {
      "/api/v1/songs": {
        "get": {
          "summary": "Gets all the songs",
          "responses": {
            "200": {
              "description": "List of SWIFT songs",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SongResponse"
                  }
                }
              }
            }
          },
          "parameters": [
            {
              "in": "query",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "description": "Song ID",
              "required": false,
              "example": "2cd0d371-79c8-4c84-abb3-6d3f58052662"
            },
            {
              "in": "query",
              "name": "song",
              "schema": {
                "type": "string"
              },
              "description": "Song name",
              "required": false,
              "example": "Dancing with Our Hands Tied"
            },
            {
              "in": "query",
              "name": "artist",
              "schema": {
                "type": "string"
              },
              "description": "artist name",
              "required": false,
              "example": "Taylor Swift"
            },
            {
              "in": "query",
              "name": "writer",
              "schema": {
                "type": "string"
              },
              "description": "writer name",
              "required": false,
              "example": "Taylor Swift"
            },
            {
              "in": "query",
              "name": "album",
              "schema": {
                "type": "string"
              },
              "description": "album name",
              "required": false,
              "example": "Reputation"
            },
            {
              "in": "query",
              "name": "year",
              "schema": {
                "type": "string"
              },
              "description": "year",
              "required": false,
              "example": 2017
            }
          ]
        }
      },
      "/api/v1/songs/song/{id}": {
        "get": {
          "summary": "Gets song by ID",
          "responses": {
            "200": {
              "description": "List of SWIFT songs",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SongResponse"
                  }
                }
              }
            }
          },
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "schema": {
                "type": "string"
              },
              "description": "Song ID",
              "required": true,
              "example": "2cd0d371-79c8-4c84-abb3-6d3f58052662"
            }
          ]
        }
      }
    },
    "components": {
      "schemas": {
        "SongResponse": {
          "type": "object",
          "required": [
            "title",
            "description",
            "price",
            "image"
          ],
          "properties": {
            "title": {
              "type": "string",
              "default": "Canon EOS 1500D DSLR Camera with 18-55mm Lens"
            },
            "description": {
              "type": "string",
              "default": "Designed for first-time DSLR owners who want impressive results straight out of the box, capture those magic moments no matter your level with the EOS 1500D. With easy to use automatic shooting modes, large 24.1 MP sensor, Canon Camera Connect app integration and built-in feature guide, EOS 1500D is always ready to go."
            },
            "price": {
              "type": "number",
              "default": 879.99
            },
            "image": {
              "type": "string",
              "default": "https://i.imgur.com/QlRphfQ.jpg"
            }
          }
        }
      }
    },
    "tags": []
  }